refactored navbar
