--- src/features/recipe/components/Heading/index.ts
@@ -1,2 +1,0 @@
-import Heading from './Heading';
-export default Heading;
