--- src/features/recipe/components/Process/index.ts
@@ -1,3 +1,0 @@
-import Process from "./Process";
-
-export default Process;
