--- src/features/recipe/components/Yields/index.ts
@@ -1,2 +1,0 @@
-import Yields from './Yields';
-export default Yields;
